Cod refactoring
